# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarRentale",
        "version": "1.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Display a listing of users",
                "operationId": "ce757d6a9dd9f3bf15fabb6611112518",
                "responses": {
                    "200": {
                        "description": "Display a listing of users"
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Store a newly created user",
                "operationId": "3c267a0bbcc2d225301918603fda0674",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Display the specified user",
                "operationId": "b3d27f655646f5454083980a35a6b4f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified user"
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Update the specified user",
                "operationId": "cc4fdc7b6760cba4a5834742ad7b4426",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Remove the specified user",
                "operationId": "8d623c822968776003e4d1009d1321b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    }
                }
            }
        },
        "/api/cars": {
            "get": {
                "tags": [
                    "cars"
                ],
                "summary": "Display a listing of cars",
                "operationId": "1fd4f1fdc0da258c5084bcdce4bbe693",
                "responses": {
                    "200": {
                        "description": "Display a listing of cars"
                    }
                }
            },
            "post": {
                "tags": [
                    "cars"
                ],
                "summary": "Store a newly created car",
                "operationId": "6f519ab2c9767d4d3b1be7cf8f1287ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mark",
                                    "model",
                                    "color",
                                    "price",
                                    "year"
                                ],
                                "properties": {
                                    "mark": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Car created successfully"
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "get": {
                "tags": [
                    "cars"
                ],
                "summary": "Display the specified car",
                "operationId": "b222bfa902eefbb6a4a875762081dcbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified car"
                    }
                }
            },
            "put": {
                "tags": [
                    "cars"
                ],
                "summary": "Update the specified car",
                "operationId": "1d232cedb5f39024dccd947cddd10087",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "mark": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Car updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "cars"
                ],
                "summary": "Remove the specified car",
                "operationId": "802b64f09e484fbb1bb3817362d30cc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car deleted successfully"
                    }
                }
            }
        },
...
    },
...
}